use size_t as datatype for taskIDs / task_id

+++ distributedcombigrid/examples/combi_example_faults/combi_example_faults.cpp
@@ -100,7 +100,7 @@ int main(int argc, char** argv) {
 
     /* create Tasks */
     TaskContainer tasks;
-    std::vector<int> taskIDs;
+    std::vector<size_t> taskIDs;
 
     for (size_t i = 0; i < levels.size(); i++) {
       //create FaultCriterion
